Refactor Authentication logic
