+++ frontend/src/containers/LoginPage/login-page.component.js
@@ -1,16 +1,16 @@
-import React, { useState } from 'react'
+import React, { useState, useContext } from 'react'
 
 import Form from 'react-bootstrap/Form'
 import Button from 'react-bootstrap/Button'
 import Container from 'react-bootstrap/Container'
 
 import axios from 'axios'
-import { useAuth } from '../../context/auth'
+import { AuthContext } from '../../context/auth'
 
 const LoginPage = ({ history }) => {
   const [username, setUsername] = useState('')
   const [password, setPassword] = useState('')
-  const { setAuthTokens } = useAuth()
+  const { setAuthTokens } = useContext(AuthContext)
 
   function validateForm () {
     return username.length > 0 && password.length > 0
@@ -29,7 +29,7 @@ const LoginPage = ({ history }) => {
 			.then(
 				res => {
   setAuthTokens(res.data)
-					// history.push('/profile')
+  history.push('/profile')
 },
 				err => {
   console.log(err)
