+++ frontend/src/helpers/getAuthTokens.js
@@ -1,0 +1,8 @@
+export const getAuthTokens = () => {
+  const raw = localStorage.getItem('tokens')
+  if (raw !== 'undefined') {
+    return JSON.parse(raw)
+  } else {
+    return null
+  }
+}
